Moved NewDocumentTypesConfig to the NewDocumentService
